fixing uploads of gzipped movie files

+++ routes/utils.js
@@ -1,17 +1,19 @@
 const fs = require("fs");
-const spawnSync = require("child_process").spawnSync
-const readChunk = require("read-chunk")
-const fileType = require("file-type")
-const AWS = require('aws-sdk');
+const spawnSync = require("child_process").spawnSync;
+const readChunk = require("read-chunk");
+const fileType = require("file-type");
+const moment = require("moment");
+const AWS = require("aws-sdk");
 var sizeOf = require("image-size");
 AWS.config.update({
-  "accessKeyId": process.env.AWS_ACCESS_KEY_S3,
-  "secretAccessKey": process.env.AWS_SECRET_ACCESS_KEY_S3,
-})
-const s3 = new AWS.S3({
-  apiVersion: '2006-03-01', region: 'us-east-2'
+  accessKeyId: process.env.AWS_ACCESS_KEY_S3,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY_S3
 });
-const winston = require('winston');
+const s3 = new AWS.S3({
+  apiVersion: "2006-03-01",
+  region: "us-east-2"
+});
+const winston = require("winston");
 
 function checkName(filePath, isVideo) {
   const filePathSplit = filePath.split("/");
@@ -73,58 +75,167 @@ function isImageFile(path) {
   let file_type = fileType(buffer)["ext"];
   return ["jpg", "png"].includes(file_type);
 }
-
+const checkBucketExists = async bucketOptions => {
+  try {
+    await s3.headBucket(bucketOptions).promise();
+    return true;
+  } catch (err) {
+    if (err.statusCode === 404) {
+      return false;
+    }
+    throw err;
+  }
+};
 const deleteBucket = async (keys, projectId, seqId, path) => {
   try {
-    let deleteKeys = []
-    for (let key of keys) {
-      deleteKeys.push({
-        Key: key
-      })
-    }
-    console.log(deleteKeys)
+    const deleteKeys = keys.map((key, i, arr) => {
+      return { Key: key };
+    });
     if (deleteKeys.length > 0) {
-      let objectData = await s3.deleteObjects({ Bucket: `ourlabels-${projectId}-${seqId}`, Delete: { Objects: deleteKeys } }).promise()
-      console.log("DELETE OBJECTS:", objectData)
+      let objectData = await s3
+        .deleteObjects({
+          Bucket: `ourlabels-${projectId}-${seqId}`,
+          Delete: { Objects: deleteKeys }
+        })
+        .promise();
+      winston.log("error", `DELETED: ${objectData.length} objects`);
     }
-    let data = await s3.deleteBucket({ Bucket: `ourlabels-${projectId}-${seqId}` }).promise()
-    console.log("DELETE:", data)
+    let bucketOptions = { Bucket: `ourlabels-${projectId}-${seqId}` };
+    const exists = await checkBucketExists(bucketOptions);
+    if (exists) {
+      let data = await s3.deleteBucket(bucketOptions).promise();
+      winston.log("error", "DELETE:" + data.length);
+    } else {
+      console.log("CANNOT DELETE:", bucketOptions);
+    }
   } catch (err) {
-    console.log("ERR4: ", err)
+    console.log("ERR4: ", err);
   }
-}
+};
 
 const listAllKeys = (token, projectId, seqId, path, accumulator, cb) => {
   var opts = { Bucket: `ourlabels-${projectId}-${seqId}` };
   if (token) opts.ContinuationToken = token;
-  s3.listObjectsV2(opts, function (err, data) {
-    let acc = accumulator.slice()
-    acc = acc.concat(data.Contents.map((datum) => { return (datum.Key) }));
-    if (data.IsTruncated)
-      listAllKeys(data.NextContinuationToken, projectId, seqId, path, acc, cb);
-    else
-      cb(acc, projectId, seqId, path)
+  s3.listObjectsV2(opts, function(err, data) {
+    if (err) {
+      // no keys? or no bucket
+      cb([], projectId, seqId, path);
+    } else {
+      // some keys
+      let acc = accumulator.slice();
+      acc = acc.concat(
+        data.Contents.map(datum => {
+          return datum.Key;
+        })
+      );
+      if (data.IsTruncated)
+        listAllKeys(
+          data.NextContinuationToken,
+          projectId,
+          seqId,
+          path,
+          acc,
+          cb
+        );
+      else cb(acc, projectId, seqId, path);
+    }
   });
-}
+};
 
-function decompressContent(filePath, newDirectory, isVideo) {
-  const tar_x_sync = spawnSync("tar", ["-jxvf", filePath, "-C", newDirectory]);
-  const files_decompressed = tar_x_sync["output"]
-    .filter(file => {
-      return file != null;
-    })
-    .map(file => {
-      let fstring = file.toString("utf8");
-      if (fstring !== "" && fstring.includes("x")) {
-        let f = fstring.substr(1).trim();
-        return f;
-      } else {
-        return fstring;
+function decompressContent(filePath, newDirectory, isVideo, mimetype) {
+  winston.log(
+    "error",
+    "line 145: fpath: " + filePath + " mime:",
+    mimetype + " newDir:" + newDirectory
+  );
+  if (mimetype === "application/gzip") {
+    winston.log("error", "in tar");
+    try {
+      try {
+        const taroutput = spawnSync(
+          "tar",
+          ["-xzf", filePath, "-C", newDirectory, "--xform='s#.*/##x'"],
+          { shell: true }
+        );
+        if (taroutput.status !== 0) {
+          winston.log("error", "not tar");
+          throw "Could not decompress with tar";
+        }
+      } catch (tarError) {
+        winston.log("error", "ERROR in tar");
+        try {
+          const gunzipoutput = spawnSync("gunzip", [filePath]);
+          if (gunzipoutput.status !== 0) {
+            winston.log("Could not decompress archive");
+            throw "Could not decompress archive";
+          }
+        } catch (gzError) {
+          throw gzError;
+        }
       }
-    })
+    } catch (allError) {
+      winston.log("error", "Got all error in gunzip:" + allError);
+      // Was gzip file but cannot decompress!
+      if (fs.existsSync(newDirectory)) {
+        spawnSync("rm", ["-rf", newDirectory]);
+      }
+      return null;
+    }
+  } else if (mimetype === "application/bzip2") {
+    try {
+      try {
+        const taroutput = spawnSync(
+          "tar",
+          ["-xjf", filePath, "-C", newDirectory, "--xform='s#.*/##x'"],
+          { shell: true }
+        );
+        if (taroutput.status !== 0) {
+          throw "Could not decompress with tar";
+        }
+      } catch (tarError) {
+        try {
+          const bzoutput = spawnSync("bunzip2", [filePath]);
+          if (bzoutput.status !== 0) {
+            throw "Could not decompress archive";
+          }
+        } catch (bzError) {
+          throw bzError;
+        }
+      }
+    } catch (allError) {
+      winston.log("error", "Got all error in bunzip2:" + allError);
+      // Was bzip2 archive but cannot decompress!
+      if (fs.existsSync(newDirectory)) {
+        spawnSync("rm", ["-rf", newDirectory]);
+      }
+      return null;
+    }
+  }
+  const ls_output = spawnSync("ls", ["-1", newDirectory]);
+  if (ls_output.status !== 0) {
+    if (fs.existsSync(newDirectory)) {
+      winston.log("error", "Error in ls on line 214");
+      spawnSync("rm", ["-rf", newDirectory]);
+    }
+    return null;
+  }
+  const files_decompressed = ls_output.output
+    .toString("utf8")
+    .replace(",", "")
+    .split("\n")
     .filter(file => {
-      return file !== "";
+      return (
+        file != "" &&
+        file != "," &&
+        !file.endsWith("gz") &&
+        !file.endsWith("bz2")
+      );
     });
+  winston.log(
+    "error",
+    "FILES EXIST FROM LS:",
+    JSON.stringify(files_decompressed)
+  );
   if (isVideo) {
     if (files_decompressed.length > 1 || files_decompressed.length === 0) {
       if (fs.existsSync(filePath)) {
@@ -140,10 +251,8 @@ function decompressContent(filePath, newDirectory, isVideo) {
     }
     let newVideoPath = `${newDirectory}/${files_decompressed[0]}`;
     if (isVideoFile(newVideoPath)) {
-      if (fs.existsSync(filePath)) {
-        fs.unlinkSync(filePath);
-      }
-      return [`${files_decompressed[0]}`];
+      winston.log("error", "found view at:" + newVideoPath);
+      return [files_decompressed[0]];
     }
   } else {
     let files_decompressed_tested = [];
@@ -182,12 +291,32 @@ function processImage(newDirectory, filename) {
   }
   return images;
 }
+const convertToHMS = secs => {
+  return (
+    moment()
+      .startOf("day")
+      .seconds(secs)
+      .format("H:mm:ss") + ".0"
+  );
+};
 
-function processVideo(newDirectory, filename, seqname) {
+const convertToMod = everyN => {
+  return `"select=not(mod(n\\,${everyN}))"`;
+};
+
+function processVideo(
+  newDirectory,
+  filename,
+  seqname,
+  mimetype,
+  beginS,
+  lengthS,
+  everyNFrames
+) {
   const filePath = `${newDirectory}/${filename}`;
   let files = [];
-  if (/.gz|.bz2/.test(filename)) {
-    files = decompressContent(filePath, newDirectory, true);
+  if (/.gz|.bz2|gzip|bzip2$/.test(filename)) {
+    files = decompressContent(filePath, newDirectory, true, mimetype);
   } else {
     if (isVideoFile(filePath)) {
       files = [filename];
@@ -195,23 +324,37 @@ function processVideo(newDirectory, filename, seqname) {
   }
   const newPath = `${newDirectory}/${files[0]}`;
   const newPrefix = `${newDirectory}/${seqname}`;
+  const begin = convertToHMS(beginS);
+  const length = convertToHMS(lengthS);
+  const everyN = convertToMod(everyNFrames);
+  winston.log("error", `FILES utils line 308: ${JSON.stringify(files)}`);
+  const ls_output = spawnSync("ls", [newDirectory]);
+  winston.log("error", ls_output.output.toString("utf8"));
   if (files != null && files.length === 1) {
     try {
       const processOptions = [
         "-i",
         newPath,
-        "-r",
-        "1/1",
-        newPrefix + "-%05d.jpg"
+        "-ss",
+        begin,
+        "-t",
+        length,
+        "-vf",
+        everyN,
+        "-vsync",
+        "vfr",
+        newPrefix + "-%06d.jpg"
       ];
-      spawnSync("ffmpeg", processOptions);
-      if (fs.existsSync(newPath)) {
-        fs.unlinkSync(newPath);
-      }
+      const ffmpeg = spawnSync("ffmpeg", processOptions, {shell:true});
+      console.log(ffmpeg.status);
+      console.log(ffmpeg.output.toString("utf8"));
       const ls = spawnSync("ls", ["-1", newDirectory]);
-      const lsString = ls.stdout.toString("utf8");
+      const lsString = ls.stdout.toString("utf8").replace(",", "");
       const images = lsString
         .split("\n")
+        .filter(file => {
+          return file != "" && file.endsWith("jpg")
+        })
         .map(file => {
           let filename = file.trim();
           let fileStats = fs.statSync(`${newDirectory}/${filename}`);
@@ -220,8 +363,10 @@ function processVideo(newDirectory, filename, seqname) {
         .filter(file => {
           return file.filename !== "";
         });
+      winston.log("error", JSON.stringify(images));
       return images;
     } catch (err) {
+      winston.log("error", "ERROR in processVideo line 355:" + err);
       if (fs.existsSync(newPath)) {
         fs.unlinkSync(newPath);
       }
@@ -229,48 +374,70 @@ function processVideo(newDirectory, filename, seqname) {
     }
   } else {
     // cancel the process
-    fs.unlinkSync(newPath);
+    if (fs.existsSync(newPath)) {
+      fs.unlinkSync(newPath);
+    }
     return [];
   }
 }
 
-const processSeqImages = async (files, seq, newDirectory, userid, projectId) => {
+const processSeqImages = async (
+  files,
+  seq,
+  newDirectory,
+  userid,
+  projectId
+) => {
   let images = [];
   for (let i = seq.begin; i <= seq.end; i += 1) {
     let processedImages = [];
     const file = files[i];
-    winston.log('error', file)
+    winston.log("error", file);
     const filename = file.filename;
     const newPath = `${newDirectory}/${filename}`;
     fs.renameSync(file.path, newPath);
     if (seq.newVideo && checkName(newPath, seq.newVideo)) {
       // process video also handles tar gz/bz2 files
-      processedImages = processVideo(newDirectory, filename, seq.newName);
+      processedImages = processVideo(
+        newDirectory,
+        filename,
+        seq.newName,
+        file.mimetype,
+        seq.newBeginS,
+        seq.newLengthS,
+        seq.newEveryNFrames
+      );
     } else if (!seq.newVideo && checkName(newPath, seq.newVideo)) {
       // process images, also handles tar gz/bz2 files
-      processedImages = processImage(newDirectory, filename, seq.newName);
+      processedImages = processImage(
+        newDirectory,
+        filename,
+        seq.newName,
+        file.mimetype
+      );
     }
     const bucketParams = {
       Bucket: `ourlabels-${projectId}-${seq.newName}`
-    }
+    };
     try {
-      console.log("BUCKET:", bucketParams)
+      winston.log("error", "BUCKET:" + JSON.stringify(bucketParams));
       try {
-        await s3.createBucket(bucketParams).promise()
+        await s3.createBucket(bucketParams).promise();
       } catch (err) {
-        console.log("Already created")
+        console.log("Already created");
       }
       for (const image of processedImages) {
         try {
-          let imageFile = fs.readFileSync(`${newDirectory}/${image.filename}`)
-          let fileDims = sizeOf(`${newDirectory}/${image.filename}`)
+          winston.log("error", "IMAGE" + JSON.stringify(image));
+          let imageFile = fs.readFileSync(`${newDirectory}/${image.filename}`);
+          let fileDims = sizeOf(`${newDirectory}/${image.filename}`);
           const fileParams = {
             Key: `${image.filename}`,
             Body: imageFile,
             Bucket: `ourlabels-${projectId}-${seq.newName}`
-          }
-          let data = await s3.upload(fileParams).promise()
-          console.log("DATA:", data)
+          };
+          let data = await s3.upload(fileParams).promise();
+          winston.log("error", "DATA:" + JSON.stringify(data));
           images.push({
             userid: userid,
             file: image.filename,
@@ -279,21 +446,19 @@ const processSeqImages = async (files, seq, newDirectory, userid, projectId) => 
             pixelWidth: fileDims.width,
             pixelHeight: fileDims.height,
             classifications: []
-          })
-          fs.unlinkSync(`${newDirectory}/${image.filename}`)
-          console.log("IMAGES:", images)
+          });
           // daily we will check if files still exist to upload
           // only delete the file if it's actually been uploaded
         } catch (err) {
-          console.log("ERR2:", err)
+          winston.log("error", `Error2: ${err}`);
         }
       }
     } catch (err) {
-      console.log("ERR3:", err)
+      winston.log("error", `Error3: ${err}`);
     }
   }
   return images;
-}
+};
 
 function passwordMeetsCriteria(password) {
   return password.length >= 16 && password.length <= 60;
@@ -304,11 +469,11 @@ function organizeClassifications(boxes, divX = 2, divY = 2) {
     boxesOrganized.push([]);
   }
   for (let box of boxes) {
-    console.log(box)
-    let rectX = Math.floor(Math.round(((box.x + box.width)/2) * divX));
-    let rectY = Math.floor(Math.round(((box.y + box.height)/2) * divY)); // both are 0 based so no -1
+    console.log(box);
+    let rectX = Math.floor(Math.round(((box.x + box.width) / 2) * divX));
+    let rectY = Math.floor(Math.round(((box.y + box.height) / 2) * divY)); // both are 0 based so no -1
     let boxPos = rectY * divX + rectX;
-    console.log(boxPos)
+    console.log(boxPos);
     boxesOrganized[boxPos].push(box);
   }
   let continuous = [];
@@ -326,7 +491,7 @@ function organizeClassifications(boxes, divX = 2, divY = 2) {
   }
   return continuous;
 }
-const generateBoxesFromBoxes = function (boxes) {
+const generateBoxesFromBoxes = function(boxes) {
   let newBoxes = [];
   for (let box of boxes) {
     newBoxes.push({
